Refactor: TimeCardView 분리
